myproject_calculator
This is a basic GUI calculator which can perform the four math operations (namely addition, subtraction, multiplication and division). It can also convert the sign of a given number (either to change to negative number or to positive number). Another feature of this program is its capability to convert a number to its decimal form once the percent button has been pressed. Whenever the user decided to clear the number that is currently being displayed at the screen, it can be cleared via the AC (All Clear) button. If in case a math operation has been made and the equal button has been pressed, the displayed result will be deleted and will be replaced by the new number entry. This calculator's dimension is fix and can be moved to any location of the computer's screen. I imported the tkinter to this program.